pyproject.toml
[tool.poetry]
name = "tap-criteo"
version = "0.0.1"
description = "`tap-criteo` is a Singer tap for Criteo, built with the Meltano SDK for Singer Taps."
authors = ["Edgar Ramírez"]
license = "Apache 2.0"

[tool.poetry.dependencies]
python = "<3.9,>=3.6.2"
requests = "^2.25.1"
singer-sdk = "^0.3.3"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
isort = "^5.9.1"
black = "^21.6b0"
mypy = "^0.910"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
multi_line_output = 3 # Vertical Hanging Indent
src_paths = "tap_criteo"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
# CLI declaration
tap-criteo = 'tap_criteo.tap:TapCriteo.cli'
